modified:   context_encoding.py 	modified:   question_encoding.py

+++ context_encoding.py
@@ -164,9 +164,9 @@ class context_bottom_up_lstm(object):
         # [num_leaves, 2*hidden_dim]
         leaf_hc = self.process_leafs(embx)
         leaf_h, leaf_c = tf.split(axis=1, num_or_size_splits=2, value=leaf_hc)
-        nodes_h = tf.identity(leaf_h)
+        nodes_h = tf.to_float(tf.identity(leaf_h))
         # [num_leaves, hidden_dim]
-        nodes_c = tf.identity(leaf_c)
+        nodes_c = tf.to_float(tf.identity(leaf_c))
         idx_var = tf.constant(0)  # tf.Variable(0,trainable=False)
         with tf.variable_scope("btp_Composition", reuse=True):
             cW = tf.get_variable("cW", [self.degree * self.hidden_dim, (self.degree + 3) * self.hidden_dim])
@@ -307,8 +307,8 @@ class context_top_down_lstm(object):
         logging.warn('get leaf embedding done')
         leaf_parent = tf.gather(tf.gather(self.t_par_leaf, idx_batch), tf.range(num_leaves))
         logging.warn('get leaf parents array done')
-        node_h = tf.identity(inodes_h)
-        node_c = tf.identity(inodes_c)
+        node_h = tf.to_float(tf.identity(inodes_h))
+        node_c = tf.to_float(tf.identity(inodes_c))
         with tf.variable_scope('td_Composition', reuse=True):
             cW = tf.get_variable('cW', [self.hidden_dim + self.emb_dim, 4 * self.hidden_dim])
             cb = tf.get_variable('cb', [4 * self.hidden_dim])
@@ -361,8 +361,8 @@ class context_top_down_lstm(object):
         root_state, root_cell = tf.split(node_states, num_or_size_splits=2, axis=0)
         root_state = tf.expand_dims(root_state, 0)
         root_cell = tf.expand_dims(root_cell, 0)
-        inode_h = tf.identity(root_state)
-        inode_c = tf.identity(root_state)
+        inode_h = tf.to_float(tf.identity(root_state))
+        inode_c = tf.to_float(tf.identity(root_state))
         idx_var = tf.constant(1)
         with tf.variable_scope('context_td_composition', reuse=True):
             cW = tf.get_variable('cW', [self.hidden_dim + self.emb_dim, 4 * self.hidden_dim])
